pyproject.toml
[tool.poetry]
name = "llm-sdg-classifier"
version = "0.1.0"
description = ""
authors = ["Finn Woelm <7606194+FinnWoelm@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
mypy = "^1.9.0"
pandas = "^2.2.1"
progress = "^1.6"
scikit-learn = "^1.4.1.post1"
tabulate = "^0.9.0"
jinja2 = "^3.1.3"
docstring-to-markdown = "^0.15"
babel = "^2.14.0"
types-tabulate = "^0.9.0.20240106"
pandas-stubs = "^2.2.1.240316"
openai = "^1.14.2"
python-dotenv = "^1.0.1"
diskcache = "^5.6.3"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.20240311"
frozendict = "^2.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
